+++ backstage-plugins/plugins/scaffolder-backend-module-harmonix/src/actions/index.ts
@@ -2,10 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export * from './create-repo-access-token';
-export * from './create-s3-bucket';
 export * from './create-secret';
 export * from './get-env-providers';
 export * from './get-component-info';
 export * from './get-ssm-parameters';
 export * from './get-platform-parameters';
 export * from './get-platform-metadata';
+export * from './write-file';